requirements.txt
torch>=2.1
torchvision>=0.16
opencv-python>=4.8
numpy>=1.24
scikit-learn>=1.3
tqdm>=4.66
pyyaml>=6.0
onnx>=1.14
onnxruntime>=1.16
onnxruntime-tools>=1.12.1
matplotlib>=3.7
